move to original useSelect to update dirtyEntityRecords
