+++ packages/editor/src/components/entities-saved-states/entity-type-list.js
@@ -9,7 +9,6 @@ import { some } from 'lodash';
 import { useSelect } from '@wordpress/data';
 import { PanelBody } from '@wordpress/components';
 import { page, layout } from '@wordpress/icons';
-import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
@@ -20,13 +19,6 @@ const ENTITY_NAME_ICONS = {
 	site: layout,
 	page,
 };
-const TRANSLATED_SITE_PROTPERTIES = {
-	title: __( 'Title' ),
-	description: __( 'Tagline' ),
-	sitelogo: __( 'Logo' ),
-	show_on_front: __( 'Show on front' ),
-	page_on_front: __( 'Page on front' ),
-};
 
 export default function EntityTypeList( {
 	list,
@@ -35,39 +27,9 @@ export default function EntityTypeList( {
 	closePanel,
 } ) {
 	const firstRecord = list[ 0 ];
-	const { entity, editList } = useSelect(
-		( select ) => {
-			const _entity = select( 'core' ).getEntity(
-				firstRecord.kind,
-				firstRecord.name
-			);
-
-			// Decouple site object into its edited pieces.
-			if ( 'root' === firstRecord.kind && 'site' === firstRecord.name ) {
-				const siteEdits = select( 'core' ).getEntityRecordEdits(
-					'root',
-					'site'
-				);
-				const _editList = [];
-				for ( const field in siteEdits ) {
-					_editList.push( {
-						kind: 'root',
-						name: 'site',
-						title: TRANSLATED_SITE_PROTPERTIES[ field ] || field,
-						key: field,
-					} );
-				}
-				return {
-					entity: _entity,
-					editList: _editList,
-				};
-			}
-
-			return {
-				entity: _entity,
-				editList: list,
-			};
-		},
+	const entity = useSelect(
+		( select ) =>
+			select( 'core' ).getEntity( firstRecord.kind, firstRecord.name ),
 		[ firstRecord.kind, firstRecord.name ]
 	);
 
@@ -77,7 +39,7 @@ export default function EntityTypeList( {
 
 	return (
 		<PanelBody title={ entity.label } initialOpen={ true } icon={ icon }>
-			{ editList.map( ( record ) => {
+			{ list.map( ( record ) => {
 				return (
 					<EntityRecordItem
 						key={ record.key || 'site' }
